Expose LLMObs content fields on guest spans
