+++ crates/temper-wasm/src/engine/guest_spans.rs
@@ -169,7 +169,23 @@ impl GuestSpanRegistry {
             tool.call_id = tracing::field::Empty,
             gen_ai.operation.name = tracing::field::Empty,
             gen_ai.provider.name = tracing::field::Empty,
+            gen_ai.system = tracing::field::Empty,
             gen_ai.request.model = tracing::field::Empty,
+            gen_ai.response.model = tracing::field::Empty,
+            gen_ai.request.temperature = tracing::field::Empty,
+            gen_ai.request.max_tokens = tracing::field::Empty,
+            gen_ai.conversation.id = tracing::field::Empty,
+            gen_ai.system_instructions = tracing::field::Empty,
+            gen_ai.input.messages = tracing::field::Empty,
+            gen_ai.prompt = tracing::field::Empty,
+            gen_ai.output.messages = tracing::field::Empty,
+            gen_ai.completion = tracing::field::Empty,
+            gen_ai.response.finish_reasons = tracing::field::Empty,
+            gen_ai.usage.input_tokens = tracing::field::Empty,
+            gen_ai.usage.output_tokens = tracing::field::Empty,
+            gen_ai.usage.cache_read_input_tokens = tracing::field::Empty,
+            gen_ai.usage.cache_creation_input_tokens = tracing::field::Empty,
+            http.response.body.size = tracing::field::Empty,
             error.type = tracing::field::Empty,
             error.message = tracing::field::Empty,
             exception.message = tracing::field::Empty,
