+++ crates/temper-wasm/src/host_trait/tests/span_hint_tests.rs
@@ -102,7 +102,23 @@ fn common_session_tool_and_llm_span_hints_are_datadog_visible_fields() {
         "tool.call_id",
         "gen_ai.operation.name",
         "gen_ai.provider.name",
+        "gen_ai.system",
         "gen_ai.request.model",
+        "gen_ai.response.model",
+        "gen_ai.request.temperature",
+        "gen_ai.request.max_tokens",
+        "gen_ai.conversation.id",
+        "gen_ai.system_instructions",
+        "gen_ai.input.messages",
+        "gen_ai.prompt",
+        "gen_ai.output.messages",
+        "gen_ai.completion",
+        "gen_ai.response.finish_reasons",
+        "gen_ai.usage.input_tokens",
+        "gen_ai.usage.output_tokens",
+        "gen_ai.usage.cache_read_input_tokens",
+        "gen_ai.usage.cache_creation_input_tokens",
+        "http.response.body.size",
     ] {
         assert_eq!(
             datadog_visible_span_hint_field(attr_key),
